requirements.txt
PyMuPDF>=1.23.0
networkx>=3.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
rank_bm25>=0.2.2
pydantic>=2.4.2
google-genai>=0.3.0
numpy<2.0.0
python-dotenv
jsonschema>=4.17.0
pytest>=7.0.0
fastapi
uvicorn
python-multipart
